requirements.txt
speechbrain
librosa
matplotlib
g2p_en
torchcontrib
datetime
lightning
loguru